requirements.txt
streamlit>=1.32.2
langchain>=0.1.16
openai>=1.30.1
python-dotenv>=1.0.1
diagrams>=0.23.4
pillow>=10.3.0
